[JXD-296] rollback
