+++ esphome/components/ota_rollback/ota_rollback.cpp
@@ -1,0 +1,5 @@
+#include "ota_rollback.h"
+
+namespace esphome {
+namespace ota {}  // namespace ota
+}  // namespace esphome
